try to use stopwatch for sync

+++ Assets/Scripts/Features/Client/IServerProvider.cs
@@ -2,7 +2,7 @@
 {
     public interface IServerProvider
     {
-        public int CurrentTimestamp { get; }
-        void SetCurrentTick(int serverTick);
+        public long CurrentTimestamp { get; }
+        void SetCurrentTick(long serverTick);
     }
 }
